Deleted defunct Card components Moved Pricing component to a component folder
